+++ PageObject.py
@@ -1,0 +1,12 @@
+class Page(object):
+    BASE_URL = 'http://sport.mail.ru/'
+    PATH = ''
+
+    def __init__(self, driver):
+        self.driver = driver
+
+    def open(self):
+        url = urlparse.urljoin(self.BASE_URL, self.PATH)
+        self.driver.get(url)
+        self.driver.maximize_window()
+
